+++ changelog.md
@@ -1,2 +1,3 @@
 - 2022/10/12: first release, support basic training test
 - 2022/10/14: support mining and infer test and all-in-one app
+- 2022/10/21: support `ymir-verifier` application